apps/llm-gateway/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.3.2
ollama==0.1.6
httpx<0.26.0,>=0.25.2
python-dotenv==1.0.0
pydantic==2.5.3
sqlalchemy
psycopg2-binary

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
nest-asyncio
